requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.2.0
azure-common==1.1.28
azure-core==1.29.5
azure-search-documents==11.4.0
azure-storage-blob==12.19.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dataclasses-json==0.6.4
distro==1.8.0
exceptiongroup==1.2.0
fastapi==0.104.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
httpx-sse==0.4.0
idna==3.5
install==1.3.5
isodate==0.6.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.14
langchain-cli==0.0.21
langchain-community==0.0.31
langchain-core==0.1.39
langchain-experimental==0.0.56
langchain-text-splitters==0.0.1
langserve==0.0.51
langsmith==0.1.39
lxml==5.2.1
markdown-it-py==3.0.0
marshmallow==3.21.1
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.3.5
orjson==3.10.0
packaging==23.2
pycparser==2.22
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
pyodbc==5.0.1
pypdf==3.17.1
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.29
sse-starlette==1.8.2
starlette==0.27.0
tenacity==8.2.3
tomlkit==0.12.4
tqdm==4.66.1
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
yarl==1.9.4
